refactor: moved all list actions related code in one component
